Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/property-escape-comment-01-expected.txt

CONSOLE MESSAGE: line 3: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-property.pl?clutter=5xyzblah&q=%22%20onload=%22alert(1)//' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
